db-auths: Split test_task file into multiple files
