+++ webserver/tests/tasks/test_post_tasks.py
@@ -1,210 +1,10 @@
 import json
-import pytest
 import re
 from unittest import mock
-from unittest.mock import Mock
 
-from datetime import datetime, timedelta
-from kubernetes.client.exceptions import ApiException
-
-from app.helpers.const import CLEANUP_AFTER_DAYS, TASK_POD_RESULTS_PATH
-from app.helpers.base_model import db
-from app.helpers.exceptions import InvalidRequest
-from app.models.task import Task
+from app.helpers.const import TASK_POD_RESULTS_PATH
 from tests.fixtures.azure_cr_fixtures import *
 from tests.fixtures.tasks_fixtures import *
-
-
-class TestGetTasks:
-    def test_get_list_tasks(
-            self,
-            client,
-            simple_admin_header
-        ):
-        """
-        Tests that admin users can see the list of tasks
-        """
-        response = client.get(
-            '/tasks/',
-            headers=simple_admin_header
-        )
-        assert response.status_code == 200
-
-    def test_get_list_tasks_base_user(
-            self,
-            client,
-            simple_user_header
-        ):
-        """
-        Tests that non-admin users cannot see the list of tasks
-        """
-        response = client.get(
-            '/tasks/',
-            headers=simple_user_header
-        )
-        assert response.status_code == 403
-
-    def test_get_task_by_id_admin(
-            self,
-            mocks_kc_tasks,
-            cr_client,
-            post_json_admin_header,
-            post_json_user_header,
-            simple_admin_header,
-            client,
-            registry_client,
-            task_body
-        ):
-        """
-        If an admin wants to check a specific task they should be allowed regardless
-        of who requested it
-        """
-        mocks_kc_tasks["tasks"].return_value.get_user_by_id.return_value = {"username": "user"}
-        resp = client.post(
-            '/tasks/',
-            data=json.dumps(task_body),
-            headers=post_json_user_header
-        )
-        assert resp.status_code == 201
-        task_id = resp.json["task_id"]
-
-        resp = client.get(
-            f'/tasks/{task_id}',
-            headers=simple_admin_header
-        )
-        assert resp.status_code == 200
-
-    @mock.patch('app.helpers.keycloak.Keycloak.is_user_admin', return_value=False)
-    @mock.patch('app.tasks_api.Keycloak.decode_token')
-    def test_get_task_by_id_non_admin_owner(
-            self,
-            mocks_decode,
-            mock_is_admin,
-            mocks_kc_tasks,
-            simple_user_header,
-            client,
-            basic_user,
-            task,
-            user_uuid
-        ):
-        """
-        If a user wants to check a specific task they should be allowed if they did request it
-        """
-        mocks_decode.return_value = {"sub": basic_user["id"]}
-        task.requested_by = basic_user["id"]
-        resp = client.get(
-            f'/tasks/{task.id}',
-            headers=simple_user_header
-        )
-        assert resp.status_code == 200, resp.json
-
-    @mock.patch('app.helpers.keycloak.Keycloak.is_user_admin', return_value=False)
-    def test_get_task_by_id_non_admin_non_owner(
-            self,
-            mock_is_admin,
-            mocks_kc_tasks,
-            simple_user_header,
-            client,
-            task
-        ):
-        """
-        If a user wants to check a specific task they should not be allowed if they did not request it
-        """
-        task_obj = db.session.get(Task, task.id)
-        task_obj.requested_by = "some random uuid"
-
-        resp = client.get(
-            f'/tasks/{task.id}',
-            headers=simple_user_header
-        )
-        assert resp.status_code == 403
-
-    def test_get_task_status_running_and_waiting(
-            self,
-            cr_client,
-            registry_client,
-            running_state,
-            waiting_state,
-            post_json_admin_header,
-            client,
-            task_body,
-            mocker,
-            task
-        ):
-        """
-        Test to verify the correct task status when it's
-        waiting or Running on k8s. Output would be similar
-        """
-        mocker.patch(
-            'app.models.task.Task.get_current_pod',
-            return_value=Mock(
-                status=Mock(
-                    container_statuses=[running_state]
-                )
-            )
-        )
-
-        response_id = client.get(
-            f'/tasks/{task.id}',
-            data=json.dumps(task_body),
-            headers=post_json_admin_header
-        )
-        assert response_id.status_code == 200
-        assert response_id.json["status"] == {'running': {'started_at': '1/1/2024'}}
-
-        mocker.patch(
-            'app.models.task.Task.get_current_pod',
-            return_value=Mock(
-                status=Mock(
-                    container_statuses=[waiting_state]
-                )
-            )
-        )
-
-        response_id = client.get(
-            f'/tasks/{task.id}',
-            data=json.dumps(task_body),
-            headers=post_json_admin_header
-        )
-        assert response_id.status_code == 200
-        assert response_id.json["status"] == {'waiting': {'started_at': '1/1/2024'}}
-
-    def test_get_task_status_terminated(
-            self,
-            terminated_state,
-            post_json_admin_header,
-            client,
-            task_body,
-            mocker,
-            task
-        ):
-        """
-        Test to verify the correct task status when it's terminated on k8s
-        """
-        mocker.patch(
-            'app.models.task.Task.get_current_pod',
-            return_value=Mock(
-                status=Mock(
-                    container_statuses=[terminated_state]
-                )
-            )
-        )
-
-        response_id = client.get(
-            f'/tasks/{task.id}',
-            data=json.dumps(task_body),
-            headers=post_json_admin_header
-        )
-        assert response_id.status_code == 200
-        expected_status = {
-            'terminated': {
-                'started_at': '1/1/2024',
-                'finished_at': '1/1/2024',
-                'reason': 'Completed successfully!',
-                'exit_code': 0
-            }
-        }
-        assert response_id.json["status"] == expected_status
 
 
 class TestPostTask:
@@ -738,228 +538,3 @@ class TestPostTask:
         assert env == "1"
         volumes = [vol.mount_path for vol in pod_body.spec.init_containers[1].volume_mounts if vol.name in ["conn-info", "keytab"]]
         assert volumes == ['/etc/krb5.conf', '/etc/principal.keytab']
-
-
-class TestCancelTask:
-    def test_cancel_task(
-            self,
-            client,
-            simple_admin_header,
-            task
-        ):
-        """
-        Test that an admin can cancel an existing task
-        """
-        response = client.post(
-            f'/tasks/{task.id}/cancel',
-            headers=simple_admin_header
-        )
-        assert response.status_code == 201
-
-    def test_cancel_404_task(
-            self,
-            client,
-            simple_admin_header
-        ):
-        """
-        Test that an admin can cancel a non-existing task returns a 404
-        """
-        response = client.post(
-            '/tasks/123456/cancel',
-            headers=simple_admin_header
-        )
-        assert response.status_code == 404
-
-
-class TestValidateTask:
-    def test_validate_task(
-            self,
-            client,
-            task_body,
-            cr_client,
-            registry_client,
-            post_json_admin_header
-        ):
-        """
-        Test the validation endpoint can be used by admins returns 201
-        """
-        response = client.post(
-            '/tasks/validate',
-            data=json.dumps(task_body),
-            headers=post_json_admin_header
-        )
-        assert response.status_code == 200
-
-    def test_validate_task_admin_missing_dataset(
-            self,
-            client,
-            task_body,
-            cr_client,
-            registry_client,
-            post_json_admin_header
-        ):
-        """
-        Test the validation endpoint can be used by admins returns
-        an error message if the dataset info is not provided
-        """
-        task_body["tags"].pop("dataset_id")
-        response = client.post(
-            '/tasks/validate',
-            data=json.dumps(task_body),
-            headers=post_json_admin_header
-        )
-        assert response.status_code == 400
-        assert response.json["error"] == "Administrators need to provide `tags.dataset_id` or `tags.dataset_name`"
-
-    def test_validate_task_basic_user(
-            self,
-            mocks_kc_tasks,
-            mocker,
-            client,
-            task_body,
-            cr_client,
-            registry_client,
-            post_json_user_header: dict[str, str],
-            access_request,
-            user_uuid,
-            dar_user
-        ):
-        """
-        Test the validation endpoint can be used by non-admins returns 201
-        """
-        mocks_kc_tasks["wrappers"].return_value.get_user_by_username.return_value = {"id": user_uuid}
-
-        post_json_user_header["project-name"] = access_request.project_name
-        response = client.post(
-            '/tasks/validate',
-            data=json.dumps(task_body),
-            headers=post_json_user_header
-        )
-        assert response.status_code == 200, response.json
-
-
-class TestTasksLogs:
-    def test_task_get_logs(
-            self,
-            post_json_admin_header,
-            client,
-            mocker,
-            terminated_state,
-            task
-        ):
-        """
-        Basic test that will allow us to return
-        the pods logs
-        """
-        mocker.patch(
-            'app.models.task.Task.get_current_pod',
-            return_value=Mock(
-                status=Mock(
-                    container_statuses=[terminated_state]
-                )
-            )
-        )
-        response_logs = client.get(
-            f'/tasks/{task.id}/logs',
-            headers=post_json_admin_header
-        )
-        assert response_logs.status_code == 200
-        assert response_logs.json["logs"] == [
-            'Example logs',
-            'another line'
-        ]
-
-    def test_task_logs_non_existent(
-            self,
-            post_json_admin_header,
-            client,
-            task
-        ):
-        """
-        Basic test that will check the appropriate error
-        is returned when the task id does not exist
-        """
-        response_logs = client.get(
-            f'/tasks/{task.id + 1}/logs',
-            headers=post_json_admin_header
-        )
-        assert response_logs.status_code == 404
-        assert response_logs.json["error"] == f"Task with id {task.id + 1} does not exist"
-
-    def test_task_waiting_get_logs(
-            self,
-            post_json_admin_header,
-            client,
-            mocker,
-            waiting_state,
-            task
-        ):
-        """
-        Basic test that will try to get logs for a pod
-        in an init state.
-        """
-        mocker.patch(
-            'app.models.task.Task.get_current_pod',
-            return_value=Mock(
-                status=Mock(
-                    container_statuses=[waiting_state]
-                )
-            )
-        )
-        response_logs = client.get(
-            f'/tasks/{task.id}/logs',
-            headers=post_json_admin_header
-        )
-        assert response_logs.status_code == 200
-        assert response_logs.json["logs"] == 'Task queued'
-
-    def test_task_not_found_get_logs(
-            self,
-            post_json_admin_header,
-            client,
-            mocker,
-            task
-        ):
-        """
-        Basic test that will try to get the logs from a missing
-        pod. This can happen if the task gets cleaned up
-        """
-        mocker.patch(
-            'app.models.task.Task.get_current_pod',
-            return_value=None
-        )
-        response_logs = client.get(
-            f'/tasks/{task.id}/logs',
-            headers=post_json_admin_header
-        )
-        assert response_logs.status_code == 400
-        assert response_logs.json["error"] == f'Task pod {task.id} not found'
-
-    def test_task_get_logs_fails(
-            self,
-            post_json_admin_header,
-            client,
-            k8s_client,
-            mocker,
-            task,
-            terminated_state
-        ):
-        """
-        Basic test that will try to get the logs, but k8s
-        will raise an ApiException. It is expected a 500 status code
-        """
-        mocker.patch(
-            'app.models.task.Task.get_current_pod',
-            return_value=Mock(
-                status=Mock(
-                    container_statuses=[terminated_state]
-                )
-            )
-        )
-        k8s_client["read_namespaced_pod_log"].side_effect = ApiException()
-        response_logs = client.get(
-            f'/tasks/{task.id}/logs',
-            headers=post_json_admin_header
-        )
-        assert response_logs.status_code == 500
-        assert response_logs.json["error"] == 'Failed to fetch the logs'
